requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
imutils>=0.5.4
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
huggingface_hub>=0.16.0